requirements-dev.txt
setuptools>=38.6.0
twine>=1.11.0
wheel>=0.31.0
tensorflow
nose
pycodestyle
coverage
Pillow
matplotlib
